Cleans fixtures structure.
