--- tests/built_ins/models/conftest.py
@@ -1,3 +1,0 @@
-'''Configuration file for models tests.
-
-'''
